Put the new classes in their own file
